feat: ESP32差分压缩固件OTA空中升级方式

+++ projects/athena-common-embedded/ground_feeling/ground_feeling.cpp
@@ -49,7 +49,8 @@ int ground_feeling_status() {
  */
 int lastTimeStatus; // 上次地感状态
 void x_task_ground_feeling_status(void *pvParameters) {
-    uint32_t chipId = get_chip_mac();
+    uint64_t chipId = get_chip_mac();
+
     while (1) {  // RTOS多任务条件： 1. 不断循环 2. 无return关键字
         // 确保上次检测是无车, 本次检测有车才上报 已上报不再上报
         int status = ground_feeling_status();
